+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Features:
 - Implement `EmulatorStart` and `EmulatorStop` for the Firestore emulator.
 - Implement the `FirebaseEmulatorService`.
 - Implement `EmulatorStart` and `EmulatorStop` for the Firebase Storage emulator.
+- Implement `EmulatorStart` and `EmulatorStop` for the Identity Platform emulator.
 
 ## v0.1.1 (2023-05-19)