+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
 - Implement `EmulatorStart` and `EmulatorStop` for the Spanner emulator.
 - Implement the `GoogleApisService`.
 - Implement the `FirebaseAppService`.
+- Implement the `GoogleAppCheckGenerateToken` function and CLI command.
 
 ## v0.1.1 (2023-05-19)